requirements.txt
pandas==2.2.3
sentence-transformers==3.2.0
duckdb==1.1.2
python-dotenv==1.0.1
docx2txt==0.8
llama-cloud==0.0.17
llama-index==0.11.8
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.1
llama-index-core==0.11.8
llama-index-embeddings-azure-openai==0.2.4
llama-index-embeddings-bedrock==0.3.0
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-anthropic==0.2.1
llama-index-llms-azure-openai==0.2.0
llama-index-llms-bedrock==0.2.1
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-chroma==0.2.0
llama-parse==0.5.5
llamaindex-py-client==0.1.19
chromadb==0.5.15
suds-py3==1.4.5.0
groq==0.9.0
requests==2.32.3
tiktoken==0.8.0
boto3==1.35.54 
geopandas==1.0.1
google-generativeai==0.8.3
openpyxl==3.1.5
xlrd==2.0.1
anthropic==0.28.1
numpy==1.26.4
tenacity==8.5.0
tqdm==4.66.5
openai==1.54.0
matplotlib==3.8.3
seaborn==0.13.2
